requirements.txt
streamlit>=1.32.0
sqlalchemy>=2.0.28
bcrypt>=4.1.2
pandas>=2.2.1
plotly>=5.19.0
reportlab>=4.1.0
google-generativeai>=0.5.2
Pillow>=10.2.0
requests>=2.31.0
python-dateutil>=2.9.0
numpy>=1.26.4
